updating to app directory pattern
