+++ apps/graphql/package.json
@@ -2,7 +2,6 @@
   "name": "@reubin/graphql",
   "version": "1.1.0",
   "private": true,
-  "type": "module",
   "scripts": {
     "start": "node dist/server",
     "build": "npx prisma generate && npm run bundle",
@@ -20,7 +19,7 @@
     "@types/node": "^18.7.20",
     "babel-jest": "^29.1.2",
     "cuid": "^2.1.8",
-    "jest": "^29.1.2",
+    "jest": "^29.2.2",
     "mercurius-integration-testing": "^6.0.0",
     "prisma": "^4.5.0",
     "rollup": "^2.79.1",
