+++ apps/ui/package.json
@@ -1,7 +1,5 @@
 {
   "name": "@reubin/ui",
-  "repository": "charliewilco/reubin.app",
-  "author": "Charlie ⚡ <charliewilc0@protonmail.com>",
   "version": "1.1.0",
   "scripts": {
     "dev": "next dev",
@@ -12,6 +10,8 @@
   },
   "dependencies": {
     "@headlessui/react": "^1.7.3",
+    "@next/env": "^13.0.0",
+    "@next/font": "^13.0.0",
     "@tailwindcss/aspect-ratio": "^0.4.2",
     "@tailwindcss/forms": "^0.5.3",
     "@tailwindcss/typography": "^0.5.7",
@@ -21,7 +21,6 @@
     "fathom-client": "^3.5.0",
     "graphql-request": "^5.0.0",
     "graphql-tag": "^2.12.6",
-    "jotai": "^1.8.6",
     "lucide-react": "^0.94.0",
     "next": "^13.0.0",
     "postcss": "^8.4.14",
@@ -33,6 +32,7 @@
     "tailwindcss": "^3.2.1",
     "typescript": "^4.8.4",
     "universal-cookie": "^4.0.4",
+    "valtio": "^1.7.4",
     "zod": "^3.19.1"
   },
   "devDependencies": {
@@ -45,8 +45,8 @@
     "@types/react-dom": "^18.0.5",
     "eslint": "^8.23.0",
     "eslint-config-next": "^13.0.0",
-    "jest": "^29.1.2",
-    "jest-environment-jsdom": "^29.1.2",
+    "jest": "^29.2.2",
+    "jest-environment-jsdom": "^29.2.2",
     "msw": "^0.47.3"
   },
   "eslintConfig": {
