+++ apps/ui/src/app/(marketing)/changelog/head.tsx
@@ -1,0 +1,10 @@
+export default function Head() {
+  return (
+    <>
+      <meta charSet="utf-8" />
+      <meta name="viewport" content="width=device-width" />
+      <title>Changelog | Reubin</title>
+      <link rel="icon" type="image/png" href="/favicon.png" />
+    </>
+  );
+}
