+++ apps/ui/src/app/(marketing)/changelog/page.tsx
@@ -1,0 +1,3 @@
+export default function ChangelogPage() {
+  return <h1>Changelog</h1>;
+}
