+++ apps/ui/src/app/(marketing)/login/head.tsx
@@ -1,0 +1,10 @@
+export default function Head() {
+  return (
+    <>
+      <title>Login | Reubin</title>
+      <meta charSet="utf-8" />
+      <meta name="viewport" content="width=device-width" />
+      <link rel="icon" type="image/png" href="/favicon.png" />
+    </>
+  );
+}
