+++ apps/ui/src/app/head.tsx
@@ -1,0 +1,10 @@
+export default function Head() {
+  return (
+    <>
+      <meta charSet="utf-8" />
+      <meta name="viewport" content="width=device-width" />
+      <title>Reubin | An RSS Client for the Next Generation</title>
+      <link rel="icon" type="image/png" href="/favicon.png" />
+    </>
+  );
+}
