+++ apps/ui/src/components/add-feed.tsx
@@ -1,3 +1,4 @@
+"use client";
 import { useCallback, useState } from "react";
 import { Plus } from "lucide-react";
