+++ apps/ui/src/components/app-header.tsx
@@ -1,4 +1,3 @@
-import { useAuthAtom } from "../hooks/useAuth";
 import { Logo } from "./logo";
 
 interface AppHeaderProps {
@@ -7,8 +6,6 @@ interface AppHeaderProps {
 }
 
 export function AppHeader(props: AppHeaderProps) {
-  const [{ token }] = useAuthAtom();
-
   return (
     <header className="w-full border-b border-zinc-200 px-4 backdrop-blur-sm backdrop-saturate-150 dark:border-zinc-700">
       <div className="flex items-center justify-between">
@@ -17,10 +14,7 @@ export function AppHeader(props: AppHeaderProps) {
           <h1 className="ml-4">{props.title}</h1>
         </div>
 
-        <div>
-          {props.children}
-          {token && <div>Logged in</div>}
-        </div>
+        <div>{props.children}</div>
       </div>
     </header>
   );
