+++ apps/ui/src/components/entry-full.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import Head from "next/head";
 import { useEffect } from "react";
 import useSWR from "swr";
