+++ apps/ui/src/components/site-footer.tsx
@@ -1,30 +1,24 @@
 import Link from "next/link";
 import { Globe, Twitter, Github } from "lucide-react";
 
+const YEAR = new Date().getFullYear();
+
 export function SiteFooter() {
   return (
     <footer className="mx-auto max-w-7xl py-2 px-2" role="contentinfo">
       <div>
         <nav className="flex flex-wrap">
           <div className="py-2 pr-5">
-            <Link href="/">
-              <a>About</a>
-            </Link>
+            <Link href="/">About</Link>
           </div>
           <div className="px-5 py-2">
-            <Link href="/changelog">
-              <a>Changelog</a>
-            </Link>
+            <Link href="/changelog">Changelog</Link>
           </div>
           <div className="px-5 py-2">
-            <Link href="/login">
-              <a>Login</a>
-            </Link>
+            <Link href="/login">Login</Link>
           </div>
           <div className="px-5 py-2">
-            <Link href="/privacy">
-              <a>Privacy Policy</a>
-            </Link>
+            <Link href="/privacy">Privacy Policy</Link>
           </div>
         </nav>
         <div className="mt-8 flex">
@@ -46,7 +40,7 @@ export function SiteFooter() {
         </div>
         <div className="mt-8">
           <p className="text-center text-sm leading-6 text-gray-400">
-            © 2021 Charlie Peters. All rights reserved.
+            © {YEAR} Charlie Peters. All rights reserved.
           </p>
         </div>
       </div>
