--- apps/ui/src/pages/changelog.tsx
@@ -1,11 +1,0 @@
-import { MarketingLayout } from "../components/layout";
-
-function Changelog() {
-  return (
-    <MarketingLayout addressbar="Changelog | Reubin">
-      <h1>Changelog</h1>
-    </MarketingLayout>
-  );
-}
-
-export default Changelog;
