+++ package.json
@@ -21,7 +21,7 @@
   },
   "devDependencies": {
     "@graphql-codegen/cli": "^2.13.7",
-    "@graphql-codegen/typescript": "^2.7.4",
+    "@graphql-codegen/typescript": "^2.8.0",
     "@graphql-codegen/typescript-graphql-request": "^4.5.6",
     "@graphql-codegen/typescript-operations": "^2.5.4",
     "@graphql-codegen/typescript-resolvers": "^2.7.4",
@@ -33,7 +33,7 @@
     "graphql-request": "^5.0.0",
     "graphql-tag": "^2.12.6",
     "husky": "^8.0.1",
-    "jest": "^29.1.2",
+    "jest": "^29.2.2",
     "prettier": "^2.7.1",
     "prettier-plugin-prisma": "^4.2.0",
     "prettier-plugin-tailwindcss": "^0.1.11",
